% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression_basic.R
\name{linear_Regression}
\alias{linear_Regression}
\title{linear_Regression}
\usage{
linear_Regression(
  y,
  x,
  intercept = T,
  CI.beta = F,
  CI.level = 0.95,
  partial.T = T,
  overall.F = T,
  r.square = F,
  Hat.matrix = F,
  GLH.F = F,
  contrast.matrix = NULL,
  c.matrix = NULL
)
}
\arguments{
\item{y}{a column vector, treated as outcome.}

\item{x}{a numerical matrix, treated as predictor.}

\item{intercept}{logical, "TRUE" by default. If "TRUE", the model includes the intercept.}

\item{CI.beta}{logical, "FALSE" by default. If "FALSE", the results will not output the confidence interval of betas.}

\item{CI.level}{a numeric number specifying the alpha-level when calculate CI.}

\item{partial.T}{logical, "TRUE" by default. If "TRUE", the results will output the results of partial T-test and corresponding p-values.}

\item{overall.F}{logical, "TRUE" by default. If "TRUE", the results will output the results of overall F-test and the corresponding p-value.}

\item{r.square}{logical, "FALSE" by default. If "FALSE", the results will not output the R Square and adjusted R Square.}

\item{Hat.matrix}{logical, "FALSE" by default. If "FALSE", the hat matrix will not be provided.}

\item{GLH.F}{logical, "FALSE" by default. If "FALSE", the function will not conduct general linear hypothesis testing.}

\item{contrast.matrix}{a numerical matrix (or vector) giving linear combinations of coefficients by rows. Only use when "GLH.F = TRUE".}

\item{c.matrix}{a vector represents the right-hand-side vector for GLH test. Only use when "GLH.F = TRUE".
Can be omitted, in which case it defaults to a vector of zeroes.}
}
\value{
A list contains regression results.
}
\description{
linear_Regression
}
\examples{
library(LinearRegression.basic)

linear_Regression(c(1, 3, 5, 7, 9), c(2, 3, 5, 8, 11))
linear_Regression(state.x77[, 4], state.x77[, c(1:3)])

}
