% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression_basic.R
\name{linear_Regression}
\alias{linear_Regression}
\title{Fitting Linear Regression Models}
\usage{
linear_Regression(
  y,
  x,
  intercept = T,
  CI.beta = F,
  CI.level = 0.95,
  partial.T = T,
  overall.F = T,
  r.square = F,
  Hat.matrix = F,
  GLH.F = F,
  contrast.matrix = NULL,
  c.matrix = NULL
)
}
\arguments{
\item{y}{a column vector, treated as outcome.}

\item{x}{a numerical matrix, treated as predictor.}

\item{intercept}{logical, "TRUE" by default.
If "FALSE", the model will not include the intercept.}

\item{CI.beta}{logical, "FALSE" by default.
If "TRUE", the results will output the confidence interval of betas.}

\item{CI.level}{a numeric number specifying the alpha-level when calculate CIs of betas (0.95 by default).
Only use when "CI.beta = TRUE"}

\item{partial.T}{logical, "TRUE" by default.
If "FALSE", the results will not output the results of partial T-test and corresponding p-values.}

\item{overall.F}{logical, "TRUE" by default.
If "FALSE", the results will not output the results of overall F-test and the corresponding p-value.}

\item{r.square}{logical, "FALSE" by default.
If "TRUE", the results will output the R Square and adjusted R Square of the fitted model.}

\item{Hat.matrix}{logical, "FALSE" by default.
If "TRUE", the hat matrix will be provided.}

\item{GLH.F}{logical, "FALSE" by default.
If "TRUE", general linear hypothesis testing will be conducted.}

\item{contrast.matrix}{a numerical matrix (or vector) giving linear combinations of coefficients by rows.
Only use when "GLH.F = TRUE".}

\item{c.matrix}{A vector represents the right-hand-side constant vector for GLH test.
It defaults to a vector of zeroes if this argument is omitted.
Only use when "GLH.F = TRUE".}
}
\value{
A list contains regression results.
}
\description{
linear_Regression is used to fit linear regression models. It can be used to carry out regression,
calculate confidence interval of interested parameters, conduct Partial T Test, Partial F Test and even
General Linear Hypothesis test.
}
\examples{
 # fit a regression model
 linear_Regression(state.x77[, 4], state.x77[, c(1:3)])

 x <- matrix(c(-1, -2, -3, -4, -5, 1, 3, 5, 7, 9), nrow = 5)
 y <- c(2, 5, 8, 11, 19)
 # fit the linear model without an intercept
 linear_Regression(y, x, intercept = FALSE)

 x <- as.matrix(state.x77[,c(1:3, 5:8)])
 y <- state.x77$`Life Exp`
 # calculate 99\% confidence interval of beta-hat
 linear_Regression(y, x, CI.beta = TRUE, CI.level = 0.99)

 # conduct GLH testing
 contrast.matrix <- matrix(c(0, 1, -1, 0, 0, 0, 0, 0), nrow = 1)
 c.matrix <- c(0)
 linear_Regression(y, x, contrast.matrix = contrast.matrix, GLH.F = TRUE)

}
